triple tap space fpr period without space
I would sometimes like a period without a space, but U last that ability
when I removed the period key. I thought that I would try a triple tap
of the space in order to get a period with no space after. Later I may
changge this to a double tap for a double tap for a period and triple
for a period and a space.

Amos King @adkron <amos@binarynoggin.com>